+++ src/plugins/weather_station/moon_phase.h
@@ -6,6 +6,6 @@
 
 #include <Arduino_compat.h>
 
-double unixTimeToJulianDate(time_t time);
-void calcMoon(time_t time, double &moonDay, uint8_t &moonPhase,char &moonPhaseFont, bool upperCase);
+float unixTimeToJulianDate(time_t time);
+void calcMoon(time_t time, float &moonDay, uint8_t &moonPhase,char &moonPhaseFont, bool upperCase);
 const __FlashStringHelper *moonPhaseName(uint8_t moonPhase);
